corrigindo tipo de dado

+++ questao08/Aluno.java
@@ -2,8 +2,8 @@ package questao08;
 
 public abstract class Aluno {
     public String nome;
-    public float nota1;
-    public float nota2;
+    public double nota1;
+    public double nota2;
 
     public Aluno(String nome) {
         this.nome = nome;
@@ -17,21 +17,21 @@ public abstract class Aluno {
         this.nome = nome;
     }
 
-    public float getNota1() {
+    public double getNota1() {
         return nota1;
     }
 
-    public void setNota1(float nota1) {
+    public void setNota1(double nota1) {
         this.nota1 = nota1;
     }
 
-    public float getNota2() {
+    public double getNota2() {
         return nota2;
     }
 
-    public void setNota2(float nota2) {
+    public void setNota2(double nota2) {
         this.nota2 = nota2;
     }
 
-    public abstract float calcularMedia();
+    public abstract double calcularMedia();
 }
